Support up to 2^32 bins

+++ include/seqan/kmer/kmer_ibf.h
@@ -179,7 +179,7 @@ public:
      * \param threads Number of threads to use.
      */
     template<typename TInt>
-    void clear(std::vector<uint16_t> const & bins, TInt&& threads)
+    void clear(std::vector<uint32_t> const & bins, TInt&& threads)
     {
         std::vector<std::future<void>> tasks;
         uint64_t chunkBlocks = filterVector.chunkSize / filterVector.blockBitSize;
@@ -208,7 +208,7 @@ public:
                     {
                         uint64_t vecPos = hashBlock * filterVector.blockBitSize;
                         uint8_t  chunkNo = vecPos / filterVector.chunkSize;
-                        for(uint16_t binNo : bins)
+                        for(uint32_t binNo : bins)
                         {
                             if (chunk == chunkNo)
                                 filterVector.unset_pos(vecPos + binNo);
@@ -230,7 +230,7 @@ public:
      * \param text Text to count occurences for.
      */
     template<typename TAnyString>
-    void select(std::vector<uint16_t> & counts, TAnyString const & text) // TODO uint16_t
+    void select(std::vector<uint32_t> & counts, TAnyString const & text)
     {
         uint16_t possible = length(text) - kmerSize + 1; // Supports text lengths up to 65535 + k
 
@@ -254,8 +254,8 @@ public:
                 hashToIndex(vecIndices[i]);
             }
 
-            uint16_t binNo = 0;
-            for (uint16_t batchNo = 0; batchNo < binWidth; ++batchNo)
+            uint32_t binNo = 0;
+            for (uint32_t batchNo = 0; batchNo < binWidth; ++batchNo)
             {
                 binNo = batchNo * intSize;
                 // get_int(idx, len) returns the integer value of the binary string of length len starting
@@ -311,9 +311,9 @@ public:
     template<typename TAnyString, typename TInt>
     inline void select(std::vector<bool> & selected, TAnyString const & text, TInt && threshold)
     {
-        std::vector<uint16_t> counts(noOfBins, 0);
+        std::vector<uint32_t> counts(noOfBins, 0);
         select(counts, text);
-        for(uint16_t binNo=0; binNo < noOfBins; ++binNo)
+        for(uint32_t binNo=0; binNo < noOfBins; ++binNo)
         {
             if(counts[binNo] >= threshold)
                 selected[binNo] = true;
